app/streamlit-chatbot-app/requirements.txt
mlflow>=3.1.0
openai
streamlit==1.44.1
databricks-sdk
python-dotenv
